- Fixed: side_t::GetLightLevel relied on the global 'linedef' variable for   automatic fake contrast. - Changed: Fake contrast now uses the WALLF_AUTOCONTRAST globally instead   of manipulating the sides' light values individually. This allows changing   the fake contrast at run time and also allows adding individual relative   lighting on top of it which is a planned UDMF feature.
SVN r991 (trunk)

+++ src/p_spec.cpp
@@ -791,12 +791,12 @@ void DWallLightTransfer::DoTransfer (BYTE lightlevel, int target, BYTE flags)
 
 		if (flags & WLF_SIDE1 && line->sidenum[0]!=NO_SIDE)
 		{
-			sides[line->sidenum[0]].Light = (BYTE)lightlevel;
+			sides[line->sidenum[0]].Light = lightlevel;
 		}
 
 		if (flags & WLF_SIDE2 && line->sidenum[1]!=NO_SIDE)
 		{
-			sides[line->sidenum[1]].Light = (BYTE)lightlevel;
+			sides[line->sidenum[1]].Light = lightlevel;
 		}
 	}
 }
